Hook up TypeScript build system through VSCode
Through the use of a tsconfig file, and the task runner (ctrl + shift +
B).

Also added a gitignore so I could ignore my node_modules because I
needed to download the TypeScript compiler. Also some rework to the code
because chordTone wasn't acting how I thought it was (was having some
NaN issues).